libc testsuite + fixes
git-svn-id: svn://kolibrios.org@6433 a494cfbc-eb01-0410-851d-a64ba20cac60

+++ programs/develop/ktcc/trunk/libc/string/strncpy.c
@@ -1,4 +1,6 @@
-char* strncpy(char* strDest,const char* strSource,int count)
+#include <string.h>
+
+char* strncpy(char* strDest,const char* strSource,size_t count)
 {
 	char* res;
 	res=strDest;
